+++ src/utils/setComments.ts
@@ -1,5 +1,5 @@
 import { workspace } from "vscode";
-import { CONFIG_COLOR, CONFIG_TOKENS } from "../extension";
+import { CONFIG_BACKUP, CONFIG_COLOR, CONFIG_TOKENS } from "../extension";
 import { triggerCommentsHide } from "./triggerCommentsHide";
 import { ExtensionService } from "../services/ExtensionService";
 
@@ -8,6 +8,7 @@ export const setComments = async (enabled: boolean) => {
   const colors = config.get<any>(CONFIG_TOKENS);
   const ext = ExtensionService.getInstance();
   const commentColor = ext.getSetting<boolean>(CONFIG_COLOR);
+  let backup = ext.getSetting<any>(CONFIG_BACKUP) || {};
 
   if (config && colors) {
     let textMateRules: any[] = colors["textMateRules"] || [];
@@ -17,6 +18,13 @@ export const setComments = async (enabled: boolean) => {
     );
 
     if (enabled) {
+      if (colors["comments"]) {
+        backup["comments"] = colors["comments"];
+      }
+      if (textMateRules.length > 0) {
+        backup["textMateRules"] = textMateRules;
+      }
+
       colors["comments"] = commentColor || "#00000000";
 
       if (commentRuleIdx >= 0) {
@@ -56,9 +64,23 @@ export const setComments = async (enabled: boolean) => {
         });
       }
     } else {
-      colors["comments"] = "";
+      if (
+        !backup?.comments &&
+        colors["comments"] &&
+        colors["comments"] !== commentColor
+      ) {
+        backup["comments"] = colors["comments"];
+      } else {
+        colors["comments"] = backup?.comments || "";
+      }
 
-      if (commentRuleIdx >= 0) {
+      if (
+        backup?.textMateRules &&
+        backup.textMateRules &&
+        backup.textMateRules?.length > 0
+      ) {
+        textMateRules = backup["textMateRules"];
+      } else if (commentRuleIdx >= 0) {
         textMateRules = textMateRules.filter(
           (r) =>
             r &&
@@ -68,10 +90,16 @@ export const setComments = async (enabled: boolean) => {
       }
     }
 
+    colors["[*Light*]"] = undefined;
+    colors["[*Dark*]"] = undefined;
     colors["textMateRules"] = textMateRules;
 
     await config.update(CONFIG_TOKENS, colors);
+    await ext.updateSetting(
+      CONFIG_BACKUP,
+      Object.keys(backup).length > 0 ? backup : undefined
+    );
 
-    triggerCommentsHide(colors && !colors["comments"]);
+    triggerCommentsHide(colors);
   }
 };